requirements.txt
bs4
telebot
urllib3
argparse
requests
cloudscraper
python-dotenv